:bulb: AnswerList Flip 기능 추가
AnswerList Cell 클릭시 해당 Cell이 Flip 되면서 뷰가 체인지 되도록 설정함
